+++ README.md
@@ -4,3 +4,5 @@
 [![Actions Status](https://github.com/IgorShayderov/frontend-project-lvl1/workflows/project-check/badge.svg)](https://github.com/IgorShayderov/frontend-project-lvl1/actions)
 
 <a href="https://codeclimate.com/github/codeclimate/codeclimate/maintainability"><img src="https://api.codeclimate.com/v1/badges/a99a88d28ad37a79dbf6/maintainability" /></a>
+
+[![asciicast](https://asciinema.org/a/g3T0Q9g6ncHX6rbKCpaYs9kaj.svg)](https://asciinema.org/a/g3T0Q9g6ncHX6rbKCpaYs9kaj)
