+++ bin/brain-games.js
@@ -1,6 +1,8 @@
 #!/usr/bin/env node
+/* eslint-disable import/no-named-as-default-member */
+/* eslint-disable import/no-named-as-default */
 import readlineSync from 'readline-sync';
-import BrainEven from './brain-even.js';
+import BrainEven from '../games/BrainEven.js';
 
 console.log('Welcome to the Brain Games!');
